........
  r683023 | josh | 2008-08-06 00:03:33 +0100 (Wed, 06 Aug 2008) | 1 line
  
  Fix JDK 1.4 compilation (after r682511 /bug 45538)
........
  r683081 | josh | 2008-08-06 02:39:44 +0100 (Wed, 06 Aug 2008) | 1 line
  
  refactoring aggregate records to a separate hierarchy. just starting
........
  r683093 | josh | 2008-08-06 04:06:18 +0100 (Wed, 06 Aug 2008) | 1 line
  
  should have been submitted with r683081
........
  r683096 | josh | 2008-08-06 04:23:10 +0100 (Wed, 06 Aug 2008) | 1 line
  
  should have been submitted with r683081
........
  r683128 | josh | 2008-08-06 07:20:59 +0100 (Wed, 06 Aug 2008) | 1 line
  
  removed TODO comment and formatted
........
  r683132 | josh | 2008-08-06 07:27:39 +0100 (Wed, 06 Aug 2008) | 1 line
  
  Partial fix for bug 45570 - Converted instance BitField fields to static 
........
  r683167 | josh | 2008-08-06 09:42:40 +0100 (Wed, 06 Aug 2008) | 1 line
  
  Added conditional format records to BiffViewer
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@683670 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/java/org/apache/poi/xssf/usermodel/XSSFSheet.java
@@ -312,15 +312,15 @@ public class XSSFSheet implements Sheet {
     	return null;
     }
 
-    public short[] getColumnBreaks() {
+    public int[] getColumnBreaks() {
         CTBreak[] brkArray = getSheetTypeColumnBreaks().getBrkArray();
         if (brkArray.length == 0) {
             return null;
         }
-        short[] breaks = new short[brkArray.length];
+        int[] breaks = new int[brkArray.length];
         for (int i = 0 ; i < brkArray.length ; i++) {
             CTBreak brk = brkArray[i];
-            breaks[i] = (short) brk.getId();
+            breaks[i] = (int)brk.getId();
         }
         return breaks;
     }
